fix(designsystem): AsisSpace の定義を int -> double に変更

+++ core/designsystem/lib/src/space/space.dart
@@ -1,13 +1,13 @@
 final class AsisSpace {
   AsisSpace._();
 
-  static const int x3s = 2;
-  static const int x2s = 4;
-  static const int xs = 6;
-  static const int s = 8;
-  static const int m = 16;
-  static const int l = 24;
-  static const int xl = 32;
-  static const int x2l = 40;
-  static const int x3l = 48;
+  static const double x3s = 2;
+  static const double x2s = 4;
+  static const double xs = 6;
+  static const double s = 8;
+  static const double m = 16;
+  static const double l = 24;
+  static const double xl = 32;
+  static const double x2l = 40;
+  static const double x3l = 48;
 }
